pyproject.toml
[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[project]
name = "omni-task-agent"
version = "0.1.0"
description = "A powerful multi-model task management system that can connect to various task management systems and help users choose and use the task management solution that best suits their needs"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "ACNet-AI", email = "agencollabnet@gmail.com"}
]
dependencies = [
    # LangGraph core dependencies
    "langgraph>=0.3.0,<0.4.0",
    "langchain-core>=0.2.38",
    # LLM providers
    "langchain-openai>=0.3.0",
    "langchain-anthropic>=0.1.1",
    # MCP related
    "langchain-mcp-adapters>=0.0.9",
    "naptha-automcp==0.1.1",
    # Other tools
    "pydantic>=2.0.0",
    "fastapi>=0.100.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.0.0",
]

[project.scripts]
ota = "omni_task_agent.cli:main"
serve_stdio = "run_mcp:serve_stdio"
serve_sse = "run_mcp:serve_sse"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[tool.setuptools]
packages = ["omni_task_agent"]

[tool.black]
line-length = 88
target-version = ["py39", "py310"]

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
markers = [
    "asyncio: marks tests as asyncio",
    "integration: marks tests that require integration (slower)",
    "unit: marks unit tests (faster)"
]
addopts = "--cov=omni_task_agent --cov-report=html --cov-report=term-missing -v"
